viz_monologue

viz_monologue

Usage

viz_monologue(
  dat,
  interpolate = TRUE,
  roll_avg = 0,
  facet = TRUE,
  annotate = TRUE
)

Arguments

dat

a dataframe with semantic distance values in columns (CosDist_) labeled by 'row_id_postsplit'

interpolate

T/F linear interpolation across missing observations of 'row_id_postsplit' (default is T)

roll_avg

window for computing a rolling average for smoothing the data, default is 0 (no smoothing)

facet

T/F facets the data by distance measure (CosDist_Glo vs. CosDist_SD15)

annotate

T/F option for annotating large semantic distance jumps z>2.5 with a red vertical line

Value

a dataframe