clean_dialogue

Cleans a transcript where there are two or more talkers. User specifies the dataframe and column name where target text is stored as arguments to the function. Default option is to lemmatize strings. Function splits and unlists text so that the output is in a one-row-per-word format marked by a unique numeric identifier (i.e., 'id_orig')

Usage

clean_dialogue(df, wordcol, whotalks, omit_stops = TRUE, lemmatize = TRUE)

Arguments

wordcol

quoted column name storing the strings that will be cleaned and split

whotalks

quoted column name with speaker/talker identities will be factorized

omit_stops

T/F user wishes to remove stopwords (default is TRUE)

lemmatize

T/F user wishes to lemmatize each string (default is TRUE)

dat

a dataframe with at least one target column of string data

Value

a dataframe one-word-per-row format with 'id_orig', 'word_clean', 'talker' vars appended